requirements.txt
flask
psycopg2-binary
pynacl
requests
python-dotenv